fthread: Add report command for runtime statistics
Add a report function and command to fthread to print runtime
statistics for all the threads that were spawned in the system.
Displayed statistics include a thread's state, spawn time, running
time, the last time it was running, and its name.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
Built fthread with CONFIG_FTHREAD_REPORT and, ran "fthread demo" and
"fthread test", and saw that it reported runtime statistics.

Change-Id: I61d73978e1700884eb8dd5df4edf10214cecf4d7
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60734
Reviewed-by: Simon Glass <sjg@chromium.org>