gen: fdt: Add a function for reading a flash map entry
A flash map describes the layout of flash memory in terms of offsets and
sizes for each region. Add a function to read a flash map entry from the
device tree.

(Unfortunately I had to change this commit to also touch cros/ but it is
a very small change and can be removed for upstreaming.)

Change-Id: I596c83d4cadac63e79da46ef167602053c7e48b8
Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chromium:244019
TEST=manual
With other changes, see that chromeos_sandbox can perform a developer
mode verified boot flow from start to finish.
Reviewed-on: https://gerrit.chromium.org/gerrit/62511
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>