gen: cros_ec: Add a function for decoding the Chrome OS EC flashmap
In order to talk to the EC properly we need to be able to understand the
layout of its internal flash memory. This permits emulation of the EC
for sandbox, and also software system with a real EC.

BUG=chromium:244019
TEST=manual
Not tested immediate as this code is not used yet
With other changes, see that chromeos_sandbox can perform a developer
mode verified boot flow from start to finish.

Change-Id: Id62d32a19522c9d286845cda781acdc4640f59b6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62512