gen: sandbox: Allow reading/writing of RAM buffer
It is useful to be able to save and restore the RAM contents of sandbox
U-Boot either for setting up tests, for later analysys, or for chaining
together multiple tests which need to keep the same memory contents.

Add a function to provide a memory file for U-Boot. This is read on
start-up and written when shutting down. If the file does not exist
on start-up, it will be created when shutting down.

BUG=chromium:244019
BRANCH=none
TEST=manual
Tested with other CLs by performing a full RW verified boot, which involves
passing the device tree in RAM from one U-Boot to the next.
Change-Id: Icb27e645d2cb7126cfde17b13e3fc6cfcea74cdd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62513