gen: sandbox: Add facility to save/restore sandbox state
It is often useful to be able to save out the state from a sandbox test
run, for analysis or to restore it later to continue a test. Add generic
infrastructure for doing this using a device tree binary file. This is
a flexible tagged file format which is already supported by U-Boot, and
it supports hierarchy if needed.

BUG=chromium:244019
BRANCH=none
TEST=manual
With other changes, see that chromeos_sandbox can perform a developer
mode verified boot flow from start to finish. With a supplied device
tree from a previous boot, it detects that the EC has not changed and
skips software sync.

Change-Id: Ida5b7e1144493a8fd16c5d197e45c2ac26927149
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62514
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>