gen: sandbox: Implement the bootm command for sandbox
When sandbox does a 'bootm' to run a kernel we cannot actually execute it.
So just exit sandbox, which is essentially what U-Boot does on other archs.
Also, allow sandbox to use bootm on any kernel, so that it can be used
to test booting of kernels from any architecture.

BUG=chromium:244019
BRANCH=none
TEST=manual
With other changes, see that chromeos_sandbox can perform a developer
mode verified boot flow from start to finish.

Change-Id: I6eedcaaaa3ee5789cf6fb331b4727b3650275517
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61954