gen: sandbox: Allow the console to work earlier
With sandbox, errors and problems may be reported before console_init_f()
is executed. For example, an argument may not parse correctly or U-Boot may
panic(). At present this output is swallowed so there is no indication what
is going wrong.

Adjust the console to deal with a very early sandbox setup, by detecting that
there is no global_data yet, and calling os functions in that case.

BUG=none
BRANCH=none
TEST=manual
Run U-Boot with an invalid memory file and see that the error message now
appears.

Change-Id: I2e60411bf3ccfeaa1b127e7a2e7b1f114be3a156
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61955