gen: sandbox: Add -i option to enter interactive mode
Normally when U-Boot starts with a command (-c option) it quits when the
command completes. Normally this is what is requires, since the test is
likely complete.

Provide an option to jump into the console instead, so that debugging or
other tasks may be performed before quitting.

BUG=none
BRANCH=none
TEST=manual
Run sandbox U-Boot with both -c and -i arguments. See that it runs the
command in -c, but then returns to the U-Boot prompt to accept more
commands.

Change-Id: I7869a91e1e82a68dcc2b230209c37506a048852d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61956