cros_ec: Correct comparison bewteen signed and unsigned numbers
Due to signed/unsigned comparison, '< sizeof(struct)' does not do the right
thing, since if ec_command() returns a -ve number we will consider this be
success.

Adjust all comparisons to avoid this problem.

This error was found with sandbox, which gives a segfault in this case. On
ARM we may instead silently fail.

We should also consider turning on -Wsign-compare to catch this sort of thing
in future.

BUG=chromium:244019
TEST=manual
Run verified boot flow on sandbox and see that it no longer segfaults when it
hits software sync (and at various other points along the way!)

(Note: b is build directory)

$ crosfw -b sandbox -V -a --force-rw
$ b/chromeos_sandbox/u-boot -d b/chromeos_sandbox/out/updated.dtb \
	-c "sb bind 0 /home/sjg/c/src/build/images/peach_pit/latest/chromiumos_test_image.bin; vboot_twostop" -\
	-spi_sf 0:0:W25Q128:b/chromeos_sandbox/out/image.bin -s x.dtb -w -r -n

Change-Id: I3a4a670c028c6dc57eda815de2e45b7c7ff8a876
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64117
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65662
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>