cros: dts: Add a FMAP region for use by dump_fmap
We found that dump_fmap is used by the firmware signer, and this requires an
FMAP. Add this in until we sort this out.

BUG=chromium:256912
BRANCH=none
TEST=manual
I can't seem to repeat the firmware updater problem, but...
Build firmware image for pit. Run 'dump_fmap' on it and see that it now works
correctly.

Change-Id: I1baf82dde3eca408830825d263638f6552754bc2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60859
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>