cros: Convert vboot_flags to linker_list
This is a cleaner method as it avoids the need for the external declarations
in the C file, and we can remove the vboot_flag_type() function altogether.

BUG=chromium:244019
BRANCH=none
TEST=manual
Use 'vbexport_test flags' command to verify that the flags still are shown
as before.

Also run a dev-mode verified boot on pit and link to check there are no
regressions.
Change-Id: Ib40da2475cd64ede002f731a1a250c2ada5cad5f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61842