fthread: Correctly set the thread last running time
The eventmanager expects a thread's lastran time to be the time at
which the thread returned control to the scheduler.  However, the
scheduler sets the lastran time to be the time the scheduler handed
control over to the thread.  This causes problems when a thread tries
to sleep because it will wake up much earlier than it expects to.  Fix
this bug by changing the scheduler to set the lastran time as the time
the thread returned control to the scheduler.  Also set the lastran
time for the scheduler, which should be the time the scheduler hands
control over to a thread.

BRANCH=none
BUG=chrome-os-partner:20721
TEST=manual
Built with later patch to report actual sleeping time and saw that
threads were sleeping at least as long as they were supposed to.

Change-Id: I2c781986485f3275632d84628250ece3108e365f
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61038
Reviewed-by: Simon Glass <sjg@chromium.org>