fthread: Add 'RUNNING' state for threads
The currently running thread does not have its state changed by the
scheduler when it gets scheduled.  This means it keeps its READY
state from when it was on the ready queue.  This can be confusing when
looking at report output because it's not clear which thread is
currently running.  Add a RUNNING state to clear up this confusion.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
Ran "fthread demo" and "fthread test" and saw that the currently
running thread's state was 'RUNNING'

Change-Id: I04b760960c8097e7ae239d3f053cc7a568ad9ece
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61041
Reviewed-by: Simon Glass <sjg@chromium.org>