cros: Support building for beaglebone from the CrOS ebuild.
The CrOS ebuild for U-Boot generally expects a config named in the
form "chromeos_$BOARD", for whatever setting of BOARD you may be
using.  This change adds such a config for beaglebone.

The new config builds u-boot.img and MLO binaries that have been
tested to work with the 3.2 kernel that's part of the Ubuntu image
used in the test lab, when combined with appropriate uEnv.txt
content.  The images haven't been tested with a kernel built from
Chrome OS sources.

BUG=None
TEST=boot the binaries on beaglebone
BRANCH=None

Change-Id: I6ef4d4ff8332931e3e83fcf36f75003202557f91
Signed-off-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61120
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>