exynos: usb: Switch USB VBUS GPIOs to be device tree configured
Some Exynos boards, such as the SMDK5250, control USB port power through
a GPIO pin. For now this had been hardcoded in the exynos5-dt board
file, but not all boards use the same pin, requiring local changes to
support different boards.

This patch moves the GPIO initialization into the USB host controller
drivers which they belong to, and uses the samsung,vbus-gpio parameter
in the device tree to configure it.

BUG=chrome-os-partner:18005
TEST=Get a Pit reworked to use the GPIO instead of the xHC's output
lines to control port power, then make sure USB still works. Also ensure
that it's still working on a Snow and on the SMDK5250.

Change-Id: I9ac12232971bb09429c3146c598720f13a5dae0b
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61197