fthread: Return actual sleep time from fthread_usleep()
When a thread calls fthread_usleep(), it is only guaranteed that it
will sleep for at least the amount of time that it requested.  Often,
it will end up sleeping for longer.  Returning the actual amount of
time slept will allow calling threads to properly deal with situations
when they sleep for longer than requested.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
Ran "fthread demo" and "fthread test" and saw that the thread sleeping
time is properly reported.

Change-Id: Ia04abae0ba66ed07cbf015fc28fdd48377051f31
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61310
Reviewed-by: Simon Glass <sjg@chromium.org>