Exynos5420: remove A7 reset flag from secondary_cpu_start()
Commit bd1283a, "Exynos5420: Add workaround for exynos iROM errata",
introduced a reset flag stored in PMU_SPARE_1 that was only used
when starting the A7s.  The issue is that the flag is initially
not set, so the first A7 to boot (CPU4) would see that is unset,
set it, enter wfi and then fail to boot.  The next A7 (CPU5) would
see that the flag is set and correctly jump to the iROM workaround
code after clearing the flag.  This results in only 2 of the A7s
being able to start at boot, CPU5 and CPU7.  Although the kernel
can work around this by setting PMU_SPARE_1 before starting any of
the A7s, let's remove the flag since it's not doing anything useful
and the logic behind it is broken.

BRANCH=none
BUG=chrome-os-partner:19820
TEST=Boot Pit.  IKS works without the kernel setting the reset flag:
localhost ~ # cat /dev/bL_status
	0 1 2 3 L2 COMMON CCI
[A15]   1 0 0 0  1    1    1
[A7]    0 1 1 1  1    1    1

Change-Id: Ief7806c60afd190344f8db0e2bd7880143c45fa0
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61316
Reviewed-by: Simon Glass <sjg@chromium.org>