fthread: Hook fthread_usleep() into udelay()
When fthread is configured to run, udelay calls should forward to the
fthread scheduler, which can choose another thread to schedule or will
go to sleep itself if there are no threads ready to be run.  This
makes it very easy to turn threading on or off since very little other
code needs to be changed.

BRANCH=none
BUG=chrome-os-partner:20946
TEST=manual

Change-Id: Ifbb503d3afd4a1d2ab6ce6e054c1a55327ea8843
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61438
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>