Exynos5420: Fix secondary core booting for thumb
When compiled SPL for Thumb secondary cores failed to boot
at the kernel boot up. Only one core came up out of 4.
This was happening because the code relocated to the
address 0x02073000 by the primary core was an ARM asm
code which was executed by the secondary cores as if it
was a thumb code.
This patch fixes the issue of secondary cores considering
relocated code as Thumb instructions and not ARM instructions
by jumping to the relocated with the help of "bx" ARM instruction.
"bx" instruction changes the 5th bit of CPSR which allows
execution unit to consider the following instructions as ARM
instructions.

BUG=chrome-os-partner:20128
TEST=Tested by booting kernel 10 times over peach (EVT1).
	Saw all 4 cores up in kernel boot log.
	With IKS patches saw switching happening
	and all 8 cores up in boot log. Also, did
	suspend-resume 50 times, it worked fine and 4 cores
	came up everytime.

Change-Id: Ib086f0db90645b9546a3edb16d9ef8743c85ec61
Signed-off-by: Akshay Saraswat <akshay.s@samsung.com>
Removed whitespace change:
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61273
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>