fthread: Track and report errors in thread sleep time
When a thread calls fthread_usleep(), it is only guaranteed that it
will sleep for _at least_ the amount of time requested.  Often it will
end up sleeping for longer.  Accumulate all the expected vs actual
sleep time differences and display them in fthread_report().

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
Booted to kernel on pit with CONFIG_FTHREAD_REPORT and saw that it
reported the average and largest error as well as the total number of
sleep calls.

Change-Id: Iee56393435662e4b5368e42497218d0590a90904
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61583
Reviewed-by: Simon Glass <sjg@chromium.org>