drop: cros: EC: Tegra: WIP - fix EC SPI comms
Added #ifdefs for Tegra to allow CROS EC SPI communication
to work. This is a temp workaround - the real fix will
be in the SPI driver to (a) drop framing/status/len bytes
(preamble) and checksum byte (end), and (b) handle a
half-duplex transaction cleanly to ignore the dummy data
bytes (0xFD) that the EC firmware puts in the RX FIFO
during the TX stage. (As per Yen Lin).

BUG=chrome-os-partner:20867
TEST=manual

Change-Id: I9a9e46add0388bfd2f8110742c31bd5771ab644d
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/61669
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>