fthread: arm: Exit fthread during cleanup
Call fthread_shutdown() while cleaning up before handing things over
to the kernel.  This guarantees that any threads that have not
terminated by this point will get to finish before the kernel is
started.

BRANCH=none
BUG=chrome-os-partner:20946
TEST=manual
Booted to the kernel with CONFIG_FTHREAD_REPORT, which is called by
fthread_shutdown() and saw that the report was displayed before the
kernel began executing.

Change-Id: Iaad129caad03e8951142db3a8eed16e3cbe42f4e
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62008
Reviewed-by: Simon Glass <sjg@chromium.org>