Tegra: SPI: Support a delay after deactivate
For devices that need some time to react after a SPI transaction
finishes, add the ability to set a delay.

Implement this as a delay on the first/next transaction to avoid
any delay in the fairly common case where a SPI transaction is
followed by other processing.

Based on: "EXYNOS: SPI: Support a delay after deactivate"

Change-Id: I9ead8a2715a56b4a418c2c3ee9fa30d806aa1452
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62126
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>