exynos: Protect eMMC boot configuration
In addition to applying write protection to the eMMC boot partitions, apply
protection to the eMMC boot configuration, which may allow switching the
partition that the eMMC boots off of to one which is not write protected.

BUG=chromium:254233
BRANCH=none
TEST=manual. On spring, boot from eMMC with the WP GPIO low, verify that the
boot configuration cannot be changed (Cannot set EXT_CSD[179] = 0x50).  Boot
partition write protection is still applied.

Change-Id: Ic9670a881698b53d51a638270af725894db80991
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62205
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>