exynos: serial: Reset FIFOs, check correct FIFO registers
The RX and TX FIFOs should be cleared of any contents before initializing
the UART.

In additional, the reference manual states that when using the FIFO, the
FIFO count and full bits should be checked to determine if there is data
to read, rather than the receive data ready bit, so switch getc and tstc
to use those bits.

BUG=chrome-os-partner:20927
BRANCH=none
TEST=UART still functional.  Data received before U-Boot loads does not affect
U-Boot's behavior.

Change-Id: Ib24f1284d616bdf7b2b310f8ec5ea408e9339890
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62218
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65220
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>