gen: usb: xhci: Handle transfer timeouts
Some USB devices (like ASIX Ethernet adapters) unfortunately think it's
a great idea to just NAK a bulk transfer forever when they don't feel
like completing it, circumventing XHCI's built-in timeout mechanism and
relying on the HCD to eventually abort the transfer. The former XHCI
code could not handle this case and ran into a BUG().

This patch adds support to gracefully abort an in-flight XHCI transfer
and uses this to recover from those cases, returning a proper error
condition to the upper layers and putting the transfer ring back into a
usable state.

BUG=chrome-os-partner:20925
TEST=Try to tftpboot from a non-responsive serverip, then unplug the
Ethernet cable. Observe how the code keeps running and can be
interrupted via Ctrl+C instead of crashing into a BUG().

Change-Id: Ifcdced566501b9e696da0ef493d7ad710d8796ca
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62307
Reviewed-by: Simon Glass <sjg@chromium.org>