Exynos5420: Make A7s boot with thumb-mode U-Boot on warm reset
On warm reset, all cores jump to the low_power_start function because iRAM
data is retained and because while executing iROM code all cores find
the jump flag 0x02020028 set. In low_power_start, cores check the reset
status and if true they clear the jump flag and jump back to 0x0.

The A7 cores do jump to 0x0 but consider following instructions as a Thumb
instructions which in turn makes them loop inside the iROM code instead of
jumping to power_down_core.

This issue is fixed by replacing the "mov pc" instruction with a "bx"
instruction which switches state along with the jump to make the execution
unit consider the branch target as an ARM instruction.

BUG=chrome-os-partner:20815
TEST=Tested warm reset 10 times and found all 8 cores
	booting 10 out of 10 times.

	[0.197898] CPU1: Booted secondary processor
	[0.215899] CPU1: thread -1, cpu 1, socket 0, mpidr 80000001
	[0.216662] CPU2: Booted secondary processor
	[0.235899] CPU2: thread -1, cpu 2, socket 0, mpidr 80000002
	[0.236623] CPU3: Booted secondary processor
	[0.255900] CPU3: thread -1, cpu 3, socket 0, mpidr 80000003
	[0.256640] CPU4: Booted secondary processor
	[0.285881] CPU4: thread -1, cpu 0, socket 1, mpidr 80000100
	[0.286641] CPU5: Booted secondary processor
	[0.315877] CPU5: thread -1, cpu 1, socket 1, mpidr 80000101
	[0.316631] CPU6: Booted secondary processor
	[0.345877] CPU6: thread -1, cpu 2, socket 1, mpidr 80000102
	[0.346657] CPU7: Booted secondary processor
	[0.375878] CPU7: thread -1, cpu 3, socket 1, mpidr 80000103
	[0.375995] Brought up 8 CPUs

Change-Id: I2ef006ef9fc853ff3b2933bc8e30f49ebd4cfccb
Signed-off-by: Akshay Saraswat <akshay.s@samsung.com>
Reformatted the commit message a little. Tested on Pit with 10 boots and 50
resumes so far. Before this commit the kernel would not even find the A7s.
Now it finds them, and successfully resumes them.
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62342