pit: cros_ec: Increase SPI speed on EC bus to 4MHz
This appears to be stable, and reduces software sync time from 4.2 sec
to 3.28 sec.

We could go farther to 10MHz, but as currently configured, we're not
meeting the setup and hold time.  It's unclear this is worth it; we
transfer 60KB of data during software sync, so increasing from 4MHz to
10MHz will save 90ms of transfer time.  That'd reduce software sync
time from 3.26 sec to 3.17 sec.

BUG=chrome-os-partner:19935
BRANCH=pit
TEST=run 'crosec test' from u-boot prompt; passes at >= 70 ns
     run 'cros_test -f swsync rw'; passes

Change-Id: I063fb6776ddc063fac5fcb5a7a0a1a8cf2148d07
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62750