fthread: Allow threads to preserve global state information
U-Boot has several global state variables, like the current i2c bus,
that affect how subsequent operations are carried out.  While this
isn't a problem in a single-threaded setting, it does become an issue
once multiple threads are involved because any thread can change the
global state that another thread depends on.  For example, if the i2c
bus is changed while the TPM thread is sleeping, TPM transactions that
are issued after the thread wakes up will be sent to the wrong bus.

This patch adds pre_start() and post_stop() functions to each thread
control block that are called right before that thread is scheduled
and right after it returns control to the scheduler, respectively.
The primary usage of these functions is to preserve and restore any
global state that is relevant to the thread that might have been
changed, either while the thread is sleeping or while it is running.
This makes it the calling code's responsibility to preserve relevant
global state rather than forcing fthread to keep track of all global
state information.

BRANCH=none
BUG=chrome-os-partner:20946
TEST=manual
Build U-Boot with CONFIG_CMD_FTHREAD, run "fthread test" and see that
the global state preservation test is successful

Change-Id: Ie84acd1703c9f5b0075e5996bd47b45f68769a51
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63074
Reviewed-by: Simon Glass <sjg@chromium.org>