fthread: arm: Correctly restore lr in THUMB mode
In THUMB mode, the context switching code restores the previously
saved lr as the new pc.  However, this behavior is inconsistent with
the non-THUMB assembly code and also doesn't correctly restore the
lr.  Change it to match the non-THUMB code.

BRANCH=none
BUG=chrome-os-partner:20946
TEST=manual
Built U-Boot with CONFIG_SYS_THUMB_BUILD and saw that the "fthread
demo" and "fthread test" commands reported successful results.

Change-Id: Ida727c718f44f12cca88d6124d003ed10f116387
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63128
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>