cros: dts: Use vanilla device trees for RW firmware
At present we use a special RW device tree for A and B firmware. This was
part of an attempt to remove the 'separate entry point hack' that Coreboot
uses when jumping to U-Boot. U-Boot considers entry through one point to
be from Coreboot and another from U-Boot, so that it knows whether to
function as RO firmware or RW.

For now we are keeping that hack, so we should remove this device tree
change. This change goes alongside this cros_bundle_firmware change:

https://gerrit.chromium.org/gerrit/#/c/62738/3

BUG=none (bug tracker down right now, will add later)
BRANCH=none
TEST=manual
Build and boot on link with the change above. See that there is no failure.

Change-Id: I3c71fdfa5ac9ca2e75e640e26b2880477e307e9f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63518
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>