exynos: lcd: Don't re-initialize lcd in RW firmware
The lcd is already initialized once in RO firmware.  Doing it again in
RW firmware is expensive because it takes ~180ms and generally looks
bad because the screen flashes on and off.

BRANCH=none
BUG=chrome-os-partner:21380
TEST=manual
Boot Pit in full RW mode and see that the lcd doesn't flash and the RW
boot time is reduced by ~180ms

Change-Id: I6175034f7feab402d802b01844dc8001479d393c
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63543
Reviewed-by: Simon Glass <sjg@chromium.org>