gen: fdt: Allow access to /chosen by other code
Add a general function to create the /chosen node and adjust fdt_support
to use it. This allows other code to add things to /chosen without
duplicating the same steps.

BUG=chromium:244019
BRANCH=none
TEST=manual
Boot through to the kernel on pit and see that the command line is still
passed correctly.

Change-Id: Ia2f20d9cdef18f6186f6c0aa505f2ff7eb4f1a56
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64111
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65663
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>