exynos: gpio: Change board strapping read to configure pins
The old code for reading the board ID strappings:
* Would require that the pin was already configured (by default) as
  an input--it wouldn't change the pinmux.
* Would restore the old state of pullups after it ran.

Change the function to configure the pins and leave them at its lowest
power state when used to read strappings.

Configuing the mux of the pins is espeically important for
exynos5250-snow where the pins share UART lines.  Other parts of
U-Boot are always configuring ALL UARTS even if they're not used by a
given board.  We should fix that other part of U-Boot eventually (it
could cause other problems), but it also is wise when we're reading
the board strappings not to rely on the default mux state of the pins.

BRANCH=none
BUG=chrome-os-partner:21481
TEST=Boot on rev3 and see rev3.0
TEST=Boot on rev4 and see rev4.0
TEST=Boot on snow PVT and see rev 2.0 (was rev 1)
TEST=Boot on snow misstuffed rev6 see rev 64.0 (was rev 6)
TEST=Boot on snow MP1.1 and see rev 3.1 (was rev 0)
TEST=On rev3.0 pit (00ZZ), see that pins are left in a state were the
last two pins have pulldowns and the first have no pull.  Look at md.l
0x13400008 1 and see "5415".  These pins are GPY7_1-4, so look and see
that bits 2-9 are %00 0001 01.
TEST=On pit, run "md.l 13400c48 1", then "gpio describe" on "gpx20"
through "gpx27".  Re-run "md.l 13400c48 1".  See no change in the
pulls.

Change-Id: I19b13dd816d71a55ef5a4615293bf4c385773732
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64286
Reviewed-on: https://gerrit.chromium.org/gerrit/65049
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>