armv7: Don't show bootstage report in SPL
Don't try to show the bootstage report when building an SPL image.
The bootstage_report() function is not defined for an SPL image and so
the compiler spits out an implicit function warning.  Fix this by
checking if we are building an SPL image.

BRANCH=none
BUG=chrome-os-partner:21662
TEST=manual
Build U-Boot with CONFIG_BOOTSTAGE_REPORT and see that the warning no
longer appears

Change-Id: Ie1468898fc0cd41a9fc559fb4aab68afc2274753
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64667
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65055
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>