exynos: dts: Set status 'disabled' for unused mmc intergace
A recent change (https://gerrit.chromium.org/gerrit/59373) caused a
failure to detect the SD card.

The problem is that the function scanning the device tree for mmc
interfaces now returns one more interface (as there is more room in
the array of node indices), rearranges the interfaces, and sticks the
last interface in the middle of the array. That last interface's node
does not have status field set to disabled, but it misses some
essential properties, which causes driver initialization to fail.

The driver probably should not be bailing out if an interface fails to
initialize, but the device tree needs to be fixed first.

BRANCH=pit
BUG=none
TEST=manual
   . u-boot detects the SD card now

Change-Id: I58eef432e1f989494e262ef953c8af0e4355f215
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65312
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65667