drop: exynos: config: Define CONFIG_POWER_TPS65090_EC for exynos5420
This change is part of a sequence of changes that works around the
fact that we don't have multiadpater i2c support in our tree and still
allows access to tps65090 on exynos5420 devices where it is behind a
tunnel.

We'll manage this by defining CONFIG_POWER_TPS65090_EC for exynos5420.

BRANCH=pit
BUG=chrome-os-partner:21326
TEST=On pit, use EC to turn off FET1 with 'i2cxfer w 0x90 0x0f 0x0e',
then press warm reset.  See that U-Boot turns FET1 back on.
TEST=Boot U-Boot on Snow and see that screen turns on.

Change-Id: I538f7e3ca6b7f31d8253f2498171a419a87222d7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65646
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>