drop: cros: exynos: Add "tps65090_ec" driver
This change is part of a sequence of changes that works around the
fact that we don't have multiadpater i2c support in our tree and still
allows access to tps65090 on exynos5420 devices where it is behind a
tunnel.

We basically copy out bits of the tps65090 driver into a "tps65090_ec"
driver (drivers/power/pmic/pmic_tps65090.c).  This is not a true PMIC
driver but implements enough tps65090 functions that the rest of the
U-Boot code can operate.  This driver simply piggybacks on the
existing cros_ec infrastructure.

BRANCH=pit
BUG=chrome-os-partner:21326
TEST=On pit, use EC to turn off FET1 with 'i2cxfer w 0x90 0x0f 0x0e',
then press warm reset.  See that U-Boot turns FET1 back on.
TEST=Boot U-Boot on Snow and see that screen turns on.
TEST=Cherry-pick I2995d54b3bf532ca4fd88beb35c0fac2f2203887 and see
reasonable messages:
  DOUG: write 0x0f => 0x0e (ret=0)
  DOUG: read 0x0f => 0x0e (ret=0)
  DOUG: write 0x14 => 0x0e (ret=0)
  DOUG: read 0x14 => 0x0e (ret=0)
  DOUG: write 0x14 => 0x0f (ret=0)
  DOUG: read 0x14 => 0x1f (ret=0)
  DOUG: write 0x0f => 0x0f (ret=0)
  DOUG: read 0x0f => 0x1f (ret=0)

Change-Id: I632c000fcf1cc0a6eadfa9252c9a151c4bf0c420
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65647