exynos: turn on FET4 (for SD card) at bootup
On all ChromeOS devices (snow, pit, ...) FET4 controls access to the
SD card VDD line.  We should make sure that U-Boot turns this power
on so that we can access the card.

Historically we haven't needed to do this because:
* On snow there's a bypass around FET4 which effectively eliminates
  it.  Even if we don't turn on FET4 the SD card is still powered.
  Turning on FET4 doesn't hurt though and is technically correct.
* On pit the EC turns on FET4 on cold bootup.

On pit we run into a problem if the kernel turns off FET4 like in
<https://gerrit.chromium.org/gerrit/#/c/65332/> and then we get a
software reset or warm reset.  In this case the EC won't know to turn
it back on.  U-Boot should do it.

BRANCH=pit
BUG=chrome-os-partner:20258
TEST=Turn off FET4 in EC console with "i2cxfer w 0x90 0x12 0x0e" and
  then do a warm reset.  Check in the EC with "i2cxfer r 0x90 " and
  see that the value is now 0x1f (FET4 is back on).
TEST=Boot kernel with <https://gerrit.chromium.org/gerrit/#/c/65332/>.
  Eject SD card and reboot.  Place SD card in.  Type in "mmc dev 1;
  mmc rescan; mmcinfo" and see that the SD card is found.
TEST=Cherry-pick I2995d54b3bf532ca4fd88beb35c0fac2f2203887 and see:
  DOUG: write 0x12 => 0x0f (ret=0)
  DOUG: read 0x12 => 0x1f (ret=0)

Change-Id: I34f03b5d1d2f43c4d1536135516635ae4479fbfe
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65650
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>