exynos: spl: separate boot mode detection into a function
Boot mode is going to be used in several places, this patch moves
existing boot mode determination code into a function.

It is a bit of a mystery to me who puts the mode into these locations,
but this code is equivalent to the existing one.

BRANCH=tot
BUG=chrome-os-partner:19927
TEST=manual
  . build PIT image, booted it off SPI and USB

Change-Id: Ib10a7876226f860b124c748f63215511a31eba57
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168511
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@gmail.com>
Tested-by: Vadim Bendebury <vbendeb@gmail.com>