exynos: spl: gen: enable dcache in SPL
When booting from SPI, SD or eMMC, enable dcache early in SPL. This
would help boot time once dcache state is handed over from SPL to
u-boot.

The size of DRAM covered by dcache is limited to 2GB, this should be
enhanced to cover actual amount of installed DRAM.

Use 16K in SRAM to store the TLB. In reserve_mmu() reuse the existing
TLB if mmu happens to be enabled.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
  . built and booted images using SD card, SPI flash and USB A-A.
    U-boot is still reinitializing the dcache in the CPU startup code,
    this is why performance improvements are not that significant.

Change-Id: I82e45a88ce6bf846c89a209469b3d7406b6fccaa
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168512