exynos: spl: Fix spl SPI driver
This change cleans up and fixes the SPL SPI driver.

The major functionality change is adding of polling after elog write
operation which ensures that the actual flash write process is
complete before resume continues.

Another improvement is dropping of initializations/deinitializations
all over the place, the only thing separating different SPI
transaction phases is toggling of CS, as it ought to be.

BRANCH=tot
BUG=chrome-os-partner:16132
TEST=manual
  . enable the test function in the end of the file, have it run for a
    few thousand iterations without errors

  . run suspend/resume test, observe it passing hundreds of times
   (there are still occasional elog failures, namely missing entries),
   not yet clear what exactly is wrong.

  . reboot a pit 50 times, observe it boot fine.

Change-Id: Ic8d927bae0f0495f6114dabb52667763254ca78d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168980
Reviewed-by: Doug Anderson <dianders@chromium.org>