cros_ec: Do not waitfor reset  unnecessary long
The very long 50ms delay after EC reset could be happening while other
u-boot activities take place. Make sure that the driver does not wait
longer than necessary before next transaction.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot through normal mode on Pit (power-refresh). All works as
before, but this saves 50ms in the RO path. This path is used in the
out-of-box experience so should be as fast as possible.

Change-Id: Id24399cbe4438d877e1942e693e316d8832d15f6
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168601
Reviewed-by: Randall Spangler <rspangler@chromium.org>