pit: reintroduce ability to build factory image
Stripped off chrome os boot images are not supposed to have certain
components in them, mainly the ability to netboot, but many other
u-boot components are not essential in the ro/rw/recovery images.

Historically setting BUILD_FACTORY_IMAGE make variable was triggering
building of a full scale image versus the stripped off one. This
ability was lost in transition between the 2011.12 and 2013.06
rebases, this patch adds it back.

The ToT by now has its own mechanism of controlling the image
composition, so this change does not have to be backported.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=none
   no actual build change yet.

Change-Id: I3b621b4abffdc19dcd6a5c54a0e12567e6ff69fa
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169125
Reviewed-by: Simon Glass <sjg@chromium.org>