exynos: smdk5420: Add SPL support for getting the board revision
The SPL cannot use the standard board_get_full_revision() since that
needs to parse the device tree.  Add a version of the function for the
SPL that uses the machine params.

Note that we allocate one of our precious INFORM3 to help us with this
task since we can't reliably read board strappings at resume time.
Previously INFORM3 wasn't used for anything.  We rely on the kernel
not to mess with this value, but that's not too crazy since we rely on
the kernel to setup INFORM0 and INFORM1 properly anyway.  Add
documentation to the U-Boot header file to document the various users
of the INFORM registers.

CQ-DEPEND=CL:173238
BRANCH=pit
BUG=chrome-os-partner:23024
TEST=suspend_stress_test (without TEST-ONLY CL).
TEST=TEST-ONLY CL that prints revision in SPL prints 5.0 on 5.0 board.
...see I54d81bdb2068b95bcaa78cf91d52daae4f51aab3 for TEST-ONLY CL.
TEST=On rev 5.0 board in modify INFORM3, suspend/resume, and see 4G:
  rh(0x10c20004) => 0x00302600
  w(0x1004080c, 0xc0de0502)
  suspend / resume
  rh(0x10c20004) => 0x00312600

Change-Id: I245273bad64bd2fc219e92f53b75aee35096a5a2
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173341