dmc: exynos5420: Gate CLKM to when reading PHY_CON13
From experiments it appears that PHY_CON13 is glitchy if we sample it
when CLKM is running.  If we stop CLKM when sampling it the glitches
all go away, so we'll do that as per Samsung suggestion.

We also check the "is it locked" bits of PHY_CON13 and loop until they
show the the value sampled actually represents a locked value.  It
doesn't appear that the glitching and "is it locked" are related, but
it seems wise to wait until the PHY tells us the value is good before
we use it.  In practice we will not loop more than a couple times (and
usually won't loop at all).

BRANCH=pit
BUG=chrome-os-partner:23556
TEST=suspend_stress_test w/ revert of 1.1V MIF

Change-Id: Ib67084548535b291435ee1486f4e9643aabb05a7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175531
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>